+++ liana/method/_liana_pipe.py
@@ -148,8 +148,9 @@ def liana_pipe(adata: anndata.AnnData,
     # Filter Resource
     resource = filter_resource(resource, adata.var_names)
 
-    print(f"Generating ligand-receptor stats for {adata.shape[0]} samples "
-          f"and {adata.shape[1]} features")
+    if verbose:
+        print(f"Generating ligand-receptor stats for {adata.shape[0]} samples "
+              f"and {adata.shape[1]} features")
 
     # Create Entities
     entities = np.union1d(np.unique(resource["ligand"]),