+++ liana/multi/__init__.py
@@ -1,1 +1,1 @@
-# Basic module to do DE between conditions
+from .to_tensor_c2c import to_tensor_c2c