+++ liana/plotting/__init__.py
@@ -1,2 +1,2 @@
-from ._dotplot import dotplot
 from ._proximity_plot import proximity_plot
+from ._dotplot import dotplot, dotplot_by_sample